requirements.txt
numpy>=1.14.2
tellurium>=2.0.12
networkx>=2.1
pygmo>=2.7
cloudpickle>=0.5.2
pymemcache>=1.4.4
toolz>=0.9.0
marshmallow>=2.15.2
arrow>=0.12.1
tornado>=5.0.2
pytest-tornado>=0.5.0
APScheduler>=3.5.1
requests>=2.18.0
yarl>=1.2.4
attrs>=18.1.0
pytest-mock>=1.10.0